vib_detecion_mini.py
Code to deploy onto the sensing device. Includes the train and test phases. The Micro:bit collects accelerometer data, computes features, calculates averages/standard-deviations, and performs z-score analysis. This is the minified version, meant to be lighter to run on the Micro:bit (as opposed to the long version, meant to be more human-readable)